refactor: separate custom widget files
