+++ src/ui/generic/widgets/progress_bar.ts
